+++ pkg/schema/qos.go
@@ -1,8 +1,8 @@
 package schema
 
 type Qos struct {
-	Name    string `json:"name"`
-	Device  string `json:"device"`
-	Ip      string `json:"ip"`
-	InSpeed int64  `json:"inSpeed"`
+	Name    string  `json:"name"`
+	Device  string  `json:"device"`
+	Ip      string  `json:"ip"`
+	InSpeed float64 `json:"inSpeed"`
 }
